# Speeding up transmission expression construction. Also allowing new lines to have 0 initial capacity.