--- a/switch_mod/trans_dispatch.py
+++ b/switch_mod/trans_dispatch.py
@@ -80,12 +80,10 @@ def define_components(mod):
 
     def LZ_TXNet_calculation(m, lz, tp):
         return (
-            sum(m.TxPowerReceived[lz_from, lz_to, t]
-                for (lz_from, lz_to, t) in m.TRANS_TIMEPOINTS
-                if lz_to == lz and t == tp) -
-            sum(m.TxPowerSent[lz_from, lz_to, t]
-                for (lz_from, lz_to, t) in m.TRANS_TIMEPOINTS
-                if lz_from == lz and t == tp))
+            sum(m.TxPowerReceived[lz_from, lz, tp] 
+                for lz_from in m.CONNECTED_LOAD_ZONES[lz]) -
+            sum(m.TxPowerSent[lz, lz_to, tp] 
+                for lz_to in m.CONNECTED_LOAD_ZONES[lz]))
     mod.LZ_TXNet = Expression(
         mod.LOAD_ZONES, mod.TIMEPOINTS,
         rule=LZ_TXNet_calculation)