-Generation of the LZ_TXNet Pyomo Expression was taking a long time, so it is replaced with a simpler expression that requires the construction of an additional Set in the trans_build module. Overall, expression construction time is reduced dramatically.
-Allowed existing Tx capacity to take a value of 0, since proposed lines do not yet exist and the model may choose or not to build them in the future. Since SWITCH takes a linear (and not binary) approach to Tx line building, then initial values of 0 are required for new lines.

I was experimenting with an input set of 1 day with hourly operations each year, for 6 periods of 5 years each (720 tps), with 190 projects, 5 possible projects, 23 load zones and 32 transmission lines. The whole solve was taking 395 s on my machine with Gurobi. By profiling it, it popped out that a surprising 230 s were being taken by the sum expressions in the LZ_TXNet_calculation Pyomo Expression construction. It seems the sum expression generator takes a lot of processing power to run through the for loop and to evaluate the if statements. So, I rewrote the expression to use a new Set so no if statements are needed. Solving time came down to 190 s with only 1.56 s being taken by the LZ_TXNet_calculation expression.

I ran another comparison with the 3_zone_toy example set and expression construction time came down from 0.00112 s to 0.00105 s, so the speedup seems to hold for smaller input sets. Also, solver time came down from 0.931 s to 0.172 s, which was interesting. Maybe this new way of building the expressions makes it easier for the solver to read them?